move metadata validation functions into util file
