+++ libs/hdf-converters/src/ckl-mapper/checklist-mapper.ts
@@ -15,15 +15,11 @@ import {
   ChecklistObject,
   ChecklistVuln,
   EmptyChecklistObject,
-  updateChecklistWithMetadata,
-  StigMetadata,
-  ChecklistMetadata
+  updateChecklistWithMetadata
 } from './checklist-jsonix-converter';
-import {Checklist, Asset, Techarea, Role, Assettype} from './checklistJsonix';
+import {Checklist, Asset} from './checklistJsonix';
 import {jsonixMapping} from './jsonixMapping';
-import {isIP, isFQDN, isMACAddress} from 'validator';
-import {Result} from '../utils/result';
-import * as Revalidator from 'revalidator';
+import {throwIfInvalidAssetMetadata} from './checklist-metadata-utils';
 
 enum ImpactMapping {
   high = 0.7,
@@ -235,125 +231,6 @@ export function getChecklistObjectFromHdf(
   return _.get(hdf, 'passthrough.checklist', EmptyChecklistObject);
 }
 
-const checklistMetadataSchema: Revalidator.JSONSchema<Asset> = {
-  properties: {
-    hostfqdn: {
-      type: 'string',
-      // STIG Viewer can autofill the FQDN as the local IP address
-      conform: (fqdn: string) => !fqdn || isFQDN(fqdn) || isIP(fqdn),
-      message: 'Host FQDN'
-    },
-    hostip: {
-      type: 'string',
-      conform: (ip: string) => !ip || isIP(ip),
-      message: 'Host IP'
-    },
-    hostmac: {
-      type: 'string',
-      conform: (mac: string) => !mac || isMACAddress(mac),
-      message: 'Host MAC'
-    },
-    role: {
-      type: 'string',
-      enum: Object.values(Role),
-      message: 'Role'
-    },
-    assettype: {
-      type: 'string',
-      enum: Object.values(Assettype),
-      message: 'Asset Type'
-    },
-    techarea: {
-      type: 'string',
-      enum: Object.values(Techarea),
-      message: 'Tech Area'
-    },
-    webordatabase: {
-      type: 'boolean',
-      message: 'Web or Database STIG'
-    }
-  }
-};
-
-const profileMetadataSchema: Revalidator.JSONSchema<StigMetadata> = {
-  properties: {
-    version: {
-      type: 'integer',
-      minimum: 0,
-      message: 'Version must be a non-negative integer'
-    },
-    releasenumber: {
-      type: 'integer',
-      minimum: 0,
-      message: 'Release number must be a non-negative integer'
-    },
-    releasedate: {
-      type: 'string',
-      conform: (date: string) => !date || !Number.isNaN(Date.parse(date)),
-      message: 'Release date must be a valid date'
-    }
-  }
-};
-
-export function validateChecklistMetadata(
-  metadata: ChecklistMetadata
-): Result<true, {invalid: string[]; message: string}> {
-  let invalid: string[] = [];
-  const messages: string[] = [];
-  const assetResult = validateChecklistAssetMetadata({
-    ...metadata,
-    webordatabase: metadata.webdbinstance === 'true',
-    targetkey: null
-  });
-  if (!assetResult.ok) {
-    invalid = invalid.concat(assetResult.error.invalid);
-    messages.push(assetResult.error.message);
-  }
-
-  for (const profile of metadata.profiles) {
-    const profileResult = validateChecklistProfileMetadata(profile);
-    if (!profileResult.ok) {
-      invalid = invalid.concat(profileResult.error.invalid);
-      messages.push(`Profile ${profile.name}: ${profileResult.error.message}`);
-    }
-  }
-
-  if (invalid.length === 0) return {ok: true, value: true};
-
-  const message = messages.join(', ');
-  return {ok: false, error: {invalid, message}};
-}
-
-export function validateChecklistAssetMetadata(
-  asset: Asset
-): Result<true, {invalid: string[]; message: string}> {
-  const errors = Revalidator.validate(asset, checklistMetadataSchema).errors;
-
-  if (errors.length === 0) return {ok: true, value: true};
-  // formats errors as: invalidField (invalidValue), otherInvalidField (otherValue), ...
-  const invalidFields = errors.map(
-    (e) => `${e.message} (${_.get(asset, e.property)})`
-  );
-  const message = `Invalid checklist metadata fields: ${invalidFields.join(', ')}`;
-  return {ok: false, error: {invalid: errors.map((e) => e.property), message}};
-}
-
-export function validateChecklistProfileMetadata(
-  metadata: StigMetadata
-): Result<true, {invalid: string[]; message: string}> {
-  const errors = Revalidator.validate(metadata, {
-    ...profileMetadataSchema
-  }).errors;
-
-  if (errors.length === 0) return {ok: true, value: true};
-  // formats errors as: invalidField (invalidValue), otherInvalidField (otherValue), ...
-  const invalidFields = errors.map(
-    (e) => `${e.message} (${_.get(metadata, e.property)})`
-  );
-  const message = `Invalid checklist profile metadata fields: ${invalidFields.join(', ')}`;
-  return {ok: false, error: {invalid: errors.map((e) => e.property), message}};
-}
-
 // baseconverter makes it difficult to assign an array to attributes using just path+transformer in this case because i think it gets instantly redirected along the 'isString' pathway due to the path pointing at a stringified json blob
 // consequently we have to use the arraytransformer, but that doesn't run if we provide a path at the top level of the object for the same reason as specified above, so we have to put the 'hdfSpecificData' object into the subobject 'data'
 // which we can then extract here
@@ -383,11 +260,6 @@ function getHdfSpecificDataAttribute(
   return data.hdfSpecificData[attribute] || undefined;
 }
 
-function throwIfInvalidMetadata(metadata: Asset) {
-  const result = validateChecklistAssetMetadata(metadata);
-  if (!result.ok) throw new Error(result.error.message);
-}
-
 /**
  * ChecklistResults is a wrapper for ChecklistMapper using the intakeType
  *  default returns a single hdf object without any modifications
@@ -411,15 +283,15 @@ export class ChecklistResults extends ChecklistJsonixConverter {
     if (typeof data === 'string') {
       this.jsonixData = super.toJsonix(data);
       this.checklistObject = super.toIntermediateObject(this.jsonixData);
-      throwIfInvalidMetadata(this.checklistObject.asset);
+      throwIfInvalidAssetMetadata(this.checklistObject.asset);
     } else if (containsChecklist(data)) {
       this.checklistObject = getChecklistObjectFromHdf(data);
-      throwIfInvalidMetadata(this.checklistObject.asset);
+      throwIfInvalidAssetMetadata(this.checklistObject.asset);
       this.jsonixData = super.fromIntermediateObject(this.checklistObject);
     } else {
       // CREATE Intermediate Object from HDF
       this.checklistObject = super.hdfToIntermediateObject(data);
-      throwIfInvalidMetadata(this.checklistObject.asset);
+      throwIfInvalidAssetMetadata(this.checklistObject.asset);
       this.jsonixData = super.fromIntermediateObject(this.checklistObject);
     }
     this.withRaw = withRaw;
